Refactored the storing of Permissions so permissions for all data instances are stored in a shared table. This follows the same pattern as Scheduling for all data instances introduced in C1 4.3.
It means one can set permissions on all data without having to implement any custom code. Its only necessary to introduce a custom SecurityEvaluator if one wants to handle Inheritance in a special way.

The IPagePermissions, IMediaFolderPermissions and IMediaFilePermissions are marked as obsolete and existing data will automatically be migrated to the new IDataPermissions type and the stores dropped at first startup.

Version bumped to 0.7.5.